+++ source/cli/parse-options.js
@@ -4,7 +4,7 @@
 */
 export default function (options)
 {
-    let { logLevel, outputDir, outputName, server, ...rest } = options;
+    let { logLevel, outputDir, outputName, server, use, ...rest } = options;
 
     if (rest.config === true) rest.config = 'acid.config.js';
 
@@ -20,6 +20,15 @@ export default function (options)
     if (outputName) output.name = outputName;
 
     if (Object.keys(output).length) rest.output = output;
+
+    if (use)
+    {
+        rest.configs = use.map(item => 
+        {
+            let [ spec, param ] = item.split('::');
+            return param ? [ spec, JSON.parse(param || null) ] : spec;
+        });
+    }
 
     if (logLevel) rest.logger = logLevel;