+++ source/service/importer/parse-file.js
@@ -11,18 +11,18 @@ let yamlRe = /\.ya?ml$/;
     @param { any }
       Parsed file contents.
     @return { Promise }
-      Resolves to the default export of the file provided.
+      Resolves to a module.
 */
 export default async function (file)
 {
     // json
     if (jsonRe.test(file))
-        return import(file, { with: { type: 'json' } }).then(mod => mod.default);
+        return import(file, { with: { type: 'json' } });
     
     // yaml
     else if (yamlRe.test(file))
-        return fs.readFile(file, 'utf8').then(data => jsyaml.load(data));
+        return fs.readFile(file, 'utf8').then(data => ({ default: jsyaml.load(data) }));
     
     // assume js
-    return import(file).then(mod => mod.default);
+    return import(file);
 }