+++ source/service/index.js
@@ -1,6 +1,3 @@
-import fs from 'node:fs'
-import parseFile from '#lib/parse-file.js'
-import { is } from '#utils'
 import controller from './controller.js'
 import getLogger from '../logging/index.js'
 
@@ -14,26 +11,14 @@ import getLogger from '../logging/index.js'
 
     If `options` is a filepath, a promise is returned.
 
-    @param { string | object } options
-      Configuration filename or configuration object.
-    @param { object } cliopts
-      Additional configuration options (usually from CLI).
-    @return { object | promise }
-      Resolves to an object with
+    @param { string | object | function } options
+      Configuration elements.
+    @return { object }
       - `run` (func): start the app
-      - `use` (func): extend the app
 */
-let service = async (options, cliopts) =>
-{
-    let serve = config => controller(config, cliopts)
-    
-    if (is.string(options) && fs.existsSync(options)) 
-        return parseFile(options).then(serve);
+let service = (...options) => controller(options, service.root)
 
-    return is.nonao(options) ? serve(options) : serve();
-}
-
-service.sync = controller;
+service.root = process.cwd();
 service.logger = config => global.log = getLogger(config)
 service.logger();